Update README.md
PBRVITS College Website

The PBRVITS college website is designed to provide students, faculty, and visitors with a seamless online experience. It offers easy access to academic resources, event updates, faculty information, and more. The site is built with a user-friendly interface, featuring responsive design for smooth navigation on both desktop and mobile devices. This project demonstrates my skills in web development, focusing on delivering a functional and visually appealing platform for educational institutions.